recipes-core: core-image_minimal.bbappend: Remove file
This was an attempt to round the size of the image up to a power of
two so that it could be used with qemu's SD card emulation. However:
a) it had a bug in the calculation (rounded to nearest not up)
b) it only works with core-image-minimal, no other image type
c) the instructions now tell the user to do this explicitly, as its
   not needed with a real SD card
So remove it entirely.