recipes-bsp: hobgoblin-bootfiles: config.txt: Use Xilinx AXI Ethernet
Previously the only Ethernet option was the Ethernet Lite, which is a low
performance non-dma device. This has now been replaced by default by the
gigabit AXI Ethernet IP from Xilinx. This is a more performant Ethernet
device, which does DMA and TCP checksum offload. However, it does
require a license to synthesize it, so the Ethernet Lite is kept as an
option.